Split out company export tests and refactor
